feat: 백엔드 Supabase 마이그레이션과 가족 동기화·응원메시지 API 추가

+++ thinq-parent-backend/src/main/java/com/example/thinq_parent/familygroup/repository/FamilyGroupMemberRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface FamilyGroupMemberRepository extends JpaRepository<FamilyGroupMember, Long> {
+public interface FamilyGroupMemberRepository extends JpaRepository<FamilyGroupMember, Integer> {
 
 	boolean existsByGroupIdAndUserId(Integer groupId, Integer userId);
